backend/requirements.txt
# Car Image AI - Backend Dependencies
python-dotenv>=1.0.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# AI / ML - Hugging Face RMBG-1.4 for background removal
# Pin to 4.x: transformers 5.x breaks BriaRMBG (all_tied_weights_keys vs _tied_weights_keys)
transformers>=4.45.0,<5.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
accelerate>=0.36.0
scikit-image>=0.21.0

# Nikon RAW (NEF) support
rawpy>=0.19.0
numpy>=1.24.0

# Optional: lighter PyTorch for CPU-only (uncomment for 4GB VPS)
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch --index-url https://download.pytorch.org/whl/cpu
